feat: remove unused code (#424)
* feat: remove unused code

* chore: fix linting issues

* chore: remove resolve ref test

* chore: update save event types

* chore: remove unused helpers

* chore: refactor has own key check
