+++ packages/live-preview-sdk/package.json
@@ -11,8 +11,7 @@
   "type": "module",
   "files": [
     "dist",
-    "react",
-    "style.css"
+    "react"
   ],
   "exports": {
     ".": {
@@ -24,10 +23,6 @@
       "import": "./dist/react.js",
       "require": "./dist/react.cjs",
       "types": "./dist/react.d.ts"
-    },
-    "./style.css": {
-      "import": "./dist/style.css",
-      "require": "./dist/style.css"
     }
   },
   "repository": {
@@ -64,7 +59,6 @@
     "contentful": "^10.3.4",
     "contentful-management": "^11.0.1",
     "graphql": "^16.6.0",
-    "graphql-tag": "^2.12.6",
     "jsdom": "^22.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
