+++ packages/live-preview-sdk/src/__tests__/index.spec.ts
@@ -1,12 +1,12 @@
 // @vitest-environment jsdom
-import { describe, it, expect, vi, Mock, afterEach, beforeAll } from 'vitest';
+import { Mock, afterEach, beforeAll, describe, expect, it, vi } from 'vitest';
 
-import { sendMessageToEditor, isInsideIframe } from '../helpers';
+import { isInsideIframe, sendMessageToEditor } from '../helpers';
 import { ContentfulLivePreview } from '../index';
 import { InspectorMode } from '../inspectorMode';
+import { InspectorModeDataAttributes } from '../inspectorMode/types';
 import { LiveUpdates } from '../liveUpdates';
 import { SaveEvent } from '../saveEvent';
-import { InspectorModeDataAttributes } from '../inspectorMode/types';
 
 vi.mock('../inspectorMode');
 vi.mock('../liveUpdates');
