+++ packages/live-preview-sdk/src/__tests__/saveEvent.spec.ts
@@ -1,9 +1,8 @@
 import { Entry } from 'contentful';
 import { Mock, afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { LivePreviewPostMessageMethods, MessageFromEditor } from '../messages';
-
 import { getAllTaggedEntries } from '../inspectorMode/utils';
+import { LivePreviewPostMessageMethods } from '../messages';
 import { SaveEvent } from '../saveEvent';
 import { ContentType } from '../types';
 
@@ -32,7 +31,7 @@ describe('SaveEvent', () => {
       method: LivePreviewPostMessageMethods.ENTRY_SAVED,
       contentType,
       entity: entry,
-    } as unknown as Omit<MessageFromEditor, 'from' | 'source'>);
+    });
 
     expect(callback).toHaveBeenCalledTimes(1);
   });
@@ -47,7 +46,7 @@ describe('SaveEvent', () => {
       method: LivePreviewPostMessageMethods.ENTRY_SAVED,
       contentType,
       entity: entry,
-    } as unknown as Omit<MessageFromEditor, 'from' | 'source'>);
+    });
 
     expect(callback).not.toHaveBeenCalled();
   });
