+++ packages/live-preview-sdk/src/constants.ts
@@ -1,4 +1,2 @@
-export const MAX_DEPTH = 10;
-
 export const LIVE_PREVIEW_EDITOR_SOURCE = 'live-preview-editor' as const;
 export const LIVE_PREVIEW_SDK_SOURCE = 'live-preview-sdk' as const;
