--- packages/live-preview-sdk/src/graphql/__tests__/fixtures/assetFromPreviewApp.json
@@ -1,11 +1,0 @@
-{
-  "__typename": "Asset",
-  "sys": {
-    "id": "bF3sZxYTHnAYJOFMNI3Vm"
-  },
-  "contentType": "image/jpeg",
-  "title": "Polarlights are marvelous",
-  "width": 9999,
-  "height": 9999,
-  "url": "https://images.ctfassets.net/c586i03gpxi9/bF3sZxYTHnAYJOFMNI3Vm/d32dd04a866df4511c701576022ac5fd/Polarlicht.jpg"
-}
