--- packages/live-preview-sdk/src/graphql/__tests__/utils.test.ts
@@ -1,11 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-import { buildCollectionName } from '../utils';
-
-describe('buildCollectionName', () => {
-  it('builds the correct name for a GraphQL collection', () => {
-    const result = buildCollectionName('topSection');
-
-    expect(result).toEqual('topSectionCollection');
-  });
-});
