--- packages/live-preview-sdk/src/graphql/index.ts
@@ -1,2 +1,0 @@
-export { updateEntry } from './entries';
-export { updateAsset } from './assets';
