--- packages/live-preview-sdk/src/helpers/__mocks__/debug.ts
@@ -1,7 +1,0 @@
-import { vi } from 'vitest';
-
-export const debug = {
-  error: vi.fn(),
-  warn: vi.fn(),
-  log: vi.fn(),
-};
