+++ packages/live-preview-sdk/src/helpers/index.ts
@@ -1,6 +1,4 @@
 export * from './debug';
-export * from './field';
+export * from './pollUrlChanges';
 export * from './utils';
 export * from './uuid';
-export * from './pollUrlChanges';
-export * from './resolveReference';
