+++ packages/live-preview-sdk/src/react.tsx
@@ -25,7 +25,7 @@ type EffectCallback = UseEffectParams[0];
 type DependencyList = UseEffectParams[1];
 type UseEffectReturn = ReturnType<typeof useEffect>;
 
-export function useDeepCompareMemoize<T>(value: T) {
+export function useDeepCompareMemoize<T>(value: T): T {
   const ref = useRef<T>(value);
   const signalRef = useRef<number>(0);
