--- packages/live-preview-sdk/src/rest/__tests__/constants.ts
@@ -1,2 +1,0 @@
-export const EN = 'en-US';
-export const referenceWithRichTextId = 'referenceWithRichTextId';
