--- packages/live-preview-sdk/src/rest/__tests__/fixtures/updateAssetFromEntryEditor.json
@@ -1,27 +1,0 @@
-{
-  "sys": {
-    "space": { "sys": { "type": "Link", "linkType": "Space", "id": "c586i03gpxi9" } },
-    "id": "2bGsPN0EuQMOCc6sx3adsn",
-    "type": "Asset",
-    "createdAt": "2023-04-11T08:52:38.857Z",
-    "updatedAt": "2023-04-13T08:45:23.626Z",
-    "environment": { "sys": { "id": "master", "type": "Link", "linkType": "Environment" } },
-    "createdBy": { "sys": { "type": "Link", "linkType": "User", "id": "5et6DcfSN7movZLX0rV8zZ" } },
-    "updatedBy": { "sys": { "type": "Link", "linkType": "User", "id": "5et6DcfSN7movZLX0rV8zZ" } },
-    "publishedCounter": 0,
-    "version": 58,
-    "automationTags": [],
-    "contentType": { "sys": { "type": "Link", "linkType": "ContentType", "id": "allFields" } }
-  },
-  "fields": {
-    "title": { "en-US": "Polarlights are marvelous!!" },
-    "description": { "en-US": "and only available in the northern parts" },
-    "file": {
-      "url": "//images.ctfassets.net/c586i03gpxi9/bF3sZxYTHnAYJOFMNI3Vm/d32dd04a866df4511c701576022ac5fd/Polarlicht_2.jpg",
-      "details": { "size": 1305911, "image": { "width": 3008, "height": 1960 } },
-      "fileName": "Polarlicht_2.jpg",
-      "contentType": "image/jpeg"
-    }
-  },
-  "metadata": { "tags": [] }
-}
