--- packages/live-preview-sdk/src/rest/__tests__/fixtures/updateFromEntryEditor.json
@@ -1,88 +1,0 @@
-{
-  "sys": {
-    "space": { "sys": { "type": "Link", "linkType": "Space", "id": "80q6jb175og9" } },
-    "id": "acPJgSOWmOM71gB62iEi4",
-    "type": "Entry",
-    "createdAt": "2023-02-22T09:00:00.556Z",
-    "updatedAt": "2023-04-05T18:39:21.564Z",
-    "environment": { "sys": { "id": "master", "type": "Link", "linkType": "Environment" } },
-    "publishedVersion": 77,
-    "publishedAt": "2023-03-29T07:45:31.872Z",
-    "firstPublishedAt": "2023-02-22T09:00:31.638Z",
-    "createdBy": { "sys": { "type": "Link", "linkType": "User", "id": "5dRQELB3imx6EN5FvaeyO8" } },
-    "updatedBy": { "sys": { "type": "Link", "linkType": "User", "id": "5dRQELB3imx6EN5FvaeyO8" } },
-    "publishedCounter": 26,
-    "version": 183,
-    "automationTags": [],
-    "contentType": { "sys": { "type": "Link", "linkType": "ContentType", "id": "allFields" } }
-  },
-  "fields": {
-    "shortText": "test test test test test ",
-    "shortTextUrl": "https://app.contentful.com",
-    "shortTextSlug": "test-page",
-    "shortTextList": ["short", "text", "list", "comma"],
-    "longTextMultiple": "long\ntext\nmultiple\nlines",
-    "richText": {
-      "data": {},
-      "content": [
-        {
-          "data": {},
-          "content": [
-            { "data": {}, "marks": [], "value": "rich text ", "nodeType": "text" },
-            {
-              "data": {},
-              "marks": [{ "type": "bold" }],
-              "value": "2 222223",
-              "nodeType": "text"
-            },
-            {
-              "data": {},
-              "marks": [{ "type": "code" }],
-              "value": "Hello World!",
-              "nodeType": "text"
-            }
-          ],
-          "nodeType": "paragraph"
-        }
-      ],
-      "nodeType": "document"
-    },
-    "numberInteger": 1,
-    "numberDecimal": 1.5,
-    "dateTime": "2023-04-01T00:00+01:00",
-    "location": { "lon": 9.67388, "lat": 36.12335 },
-    "mediaOneFile": {
-      "sys": { "type": "Link", "linkType": "Asset", "id": "2bGsPN0EuQMOCc6sx3adsn" }
-    },
-    "mediaManyFiles": [
-      { "sys": { "type": "Link", "linkType": "Asset", "id": "2bGsPN0EuQMOCc6sx3adsn" } },
-      { "sys": { "type": "Link", "linkType": "Asset", "id": "4YRUBbjYKnVAU17retktzF" } }
-    ],
-    "boolean": false,
-    "json": { "Locale": "en-US" },
-    "refOneSameSpace": {
-      "sys": { "type": "Link", "linkType": "Entry", "id": "2wx02IajQQMYggaxBT8CF3" }
-    },
-    "refManySameSpace": [
-      { "sys": { "type": "Link", "linkType": "Entry", "id": "422a9DU6Yb01C8Hh9ISq5c" } },
-      { "sys": { "type": "Link", "linkType": "Entry", "id": "1Gj746GwfYgQwFL0rcjmhE" } }
-    ],
-    "refOneDiffSpace": {
-      "sys": {
-        "type": "ResourceLink",
-        "linkType": "Contentful:Entry",
-        "urn": "crn:contentful:::content:spaces/3yzzz65gh9tw/entries/1oU53EU4UJjLk73abzyZV3"
-      }
-    },
-    "refManyDiffSpace": [
-      {
-        "sys": {
-          "type": "ResourceLink",
-          "linkType": "Contentful:Entry",
-          "urn": "crn:contentful:::content:spaces/3yzzz65gh9tw/entries/1oU53EU4UJjLk73abzyZV3"
-        }
-      }
-    ]
-  },
-  "metadata": { "tags": [] }
-}
