--- packages/live-preview-sdk/src/rest/__tests__/utils.ts
@@ -1,15 +1,0 @@
-import type { Asset, Entry } from 'contentful';
-
-export function patchField<T extends Asset | Entry>(
-  originalData: T,
-  name: keyof T['fields'],
-  value: unknown
-) {
-  return {
-    ...originalData,
-    fields: {
-      ...originalData.fields,
-      [name]: value,
-    },
-  };
-}
