--- packages/live-preview-sdk/src/rest/index.ts
@@ -1,2 +1,0 @@
-export { updateEntity } from './entities';
-export type { Reference } from './entities';
