+++ packages/live-preview-sdk/src/saveEvent.ts
@@ -34,10 +34,11 @@ export class SaveEvent {
     };
   }
 
-  public receiveMessage(message: Omit<MessageFromEditor, 'from' | 'source'>): void {
+  public receiveMessage(
+    message: Omit<MessageFromEditor, 'from' | 'source'> | EntrySavedMessage
+  ): void {
     if (message.method === LivePreviewPostMessageMethods.ENTRY_SAVED && this.subscription) {
-      // @TODO: will fix this type as part of this ticket https://contentful.atlassian.net/browse/TOL-1599
-      const { entity } = message as unknown as EntrySavedMessage;
+      const { entity } = message as EntrySavedMessage;
       const entries = getAllTaggedEntries();
 
       if (entries.includes(entity.sys.id)) {
