--- packages/live-preview-sdk/src/styles.css
@@ -1,1 +1,0 @@
-/* TODO: remove with next breaking version */
