--- packages/live-preview-sdk/style.css/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@contentful/live-preview/react/style.css",
-  "version": "0.0.0-determined-by-semantic-release",
-  "author": "Contentful GmbH",
-  "license": "MIT",
-  "description": "Preview SDK for both the field tagging connection + live content updates (style)",
-  "source": "../src/style.css",
-  "main": "../dist/style.css",
-  "style": "../dist/style.css",
-  "files": [
-    "../dist/style.css"
-  ]
-}
